tests/submission/PhilipTaylor/canvas/size.attributes.get.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: size.attributes.get</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.size.html">size</a>.<a href="index.size.attributes.html">attributes</a>.get</h1>
<p class="desc">Getting width/height IDL/content attributes</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.size.width">size.width</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.size.height">size.height</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="60" id="c" width="120"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="size.attributes.get.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Getting width/height IDL/content attributes");
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 120, "canvas.width", "120");
_assertSame(canvas.height, 60, "canvas.height", "60");
_assertSame(canvas.getAttribute('width'), '120', "canvas.getAttribute('width')", "'120'");
_assertSame(canvas.getAttribute('height'), '60', "canvas.getAttribute('height')", "'60'");


});
</script>

<div id="log"></div>
</body></html>
